refactor : Field Manager 제거
