--- Assets/Script/Manager/FieldManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eeb043a8439249b439f1bdc4ba12d592
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
